messed up. fixing export

+++ src/exec/exec.c
@@ -6,7 +6,7 @@
 /*   By: daniego2 <daniego2@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2025/01/10 16:59:19 by daniego2          #+#    #+#             */
-/*   Updated: 2025/04/03 19:57:56 by daniego2         ###   ########.fr       */
+/*   Updated: 2025/04/07 20:18:16 by daniego2         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -34,27 +34,19 @@ int	create_fork(t_cmd *token, char *path, t_env *env, int *standard_input, int e
 	if (pid == 0)
 	{
 		if (token->out_fd != -1)
-		{
 			dup2(token->out_fd, STDOUT_FILENO);
-		}
 		else if (token->next != NULL)
-		{
 			dup2(fd[1], STDOUT_FILENO);
-		}
 		if (token->in_fd != -1)
-		{
 			dup2(token->in_fd, STDIN_FILENO);
-		}
 		else if (*standard_input != STDIN_FILENO)
-		{
 			dup2(*standard_input, STDIN_FILENO);
-		}
 		close(fd[0]);
 		close(fd[1]);
 		if (is_builtin(token->command[0]))
 		{
-			write (1, "Executing builtin\n", 18);
-			exec_builtin(token->command, env);
+			exit_status = exec_builtin(token->command, env);
+			exit(0);
 		}
 		else
 			execve(path, token->command, assemble_env(env));
@@ -92,15 +84,26 @@ int exec(t_env *env, t_cmd *token)
 
 	exit_status = 0;
 	standard_input = STDIN_FILENO;
+	if (ft_strncmp(token->command[0], "salida", 6) == 0)
+	{
+		printf("Exit status: %d\n", exit_status);
+		return (exit_status);
+	}
 	while (token != NULL)
 	{
 		token->in_fd = get_in_fd(token);
 		token->out_fd = get_out_fd(token);
 		path_batch = get_path(assemble_env(env), path);
-		path = path_finder(path_batch, token->command[0], token);
-		printf("Path: %s\n", path);
+		path = path_finder(path_batch, token->command[0]);
+		if (!path && !is_builtin(token->command[0]))
+		{
+			free(path_batch);
+			printf("Error: No hay Path maquinón\n");
+			return (exit_status);
+		}
+		//printf("Path: %s\n", path);
 		exit_status = create_fork(token, path, env, &standard_input, exit_status);
-		printf("Sale del fork\n");
+		//printf("Sale del fork\n");
 			unlink(".here_doc");
 		token = token->next;
 	}
